perf(www): lazy-load the docs search dialog
The header mounts SearchCommand on every page, statically importing the
Orama static client, fumadocs' useDocsSearch and react-aria's
Autocomplete (~25 KB gz) even though search only opens on ⌘K.

Split the heavy body into search-dialog.tsx behind React.lazy — the
trigger, keyboard shortcut and modal/drawer shell stay eager. Hover or
focus on the trigger warms the chunk so first open feels instant.
